fix: PR Fixes - rename variants-result-provider to result-variants-provider
EX-6762
