+++ packages/x-components/src/components/result/__tests__/result-variants-provider-and-selector.spec.ts
@@ -2,7 +2,7 @@ import { mount, Wrapper, WrapperArray } from '@vue/test-utils';
 import { Result } from '@empathyco/x-types';
 import { createResultStub } from '../../../__stubs__/index';
 import { findTestDataById, getDataTestSelector, installNewXPlugin } from '../../../__tests__/utils';
-import VariantsResultProvider from '../variants-result-provider.vue';
+import ResultVariantsProvider from '../result-variants-provider.vue';
 import ResultVariantSelector from '../result-variant-selector.vue';
 import { XPlugin } from '../../../plugins/index';
 
@@ -45,19 +45,19 @@ const result = createResultStub('jacket', {
   variants
 });
 
-const renderVariantsResultProvider = ({
+const renderResultVariantsProvider = ({
   template = '<ResultVariantSelector/>',
   result
-}: VariantsResultProviderOptions): VariantsResultProviderApi => {
+}: ResultVariantsProviderOptions): ResultVariantsProviderApi => {
   const [, localVue] = installNewXPlugin();
 
   const wrapper = mount(
     {
-      template: `<VariantsResultProvider :result="result" #default="{ result: newResult }">
+      template: `<ResultVariantsProvider :result="result" #default="{ result: newResult }">
           ${template}
-        </VariantsResultProvider>`,
+        </ResultVariantsProvider>`,
       components: {
-        VariantsResultProvider,
+        ResultVariantsProvider,
         ResultVariantSelector
       }
     },
@@ -77,7 +77,7 @@ const renderVariantsResultProvider = ({
   );
 
   return {
-    wrapper: wrapper.findComponent(VariantsResultProvider),
+    wrapper: wrapper.findComponent(ResultVariantsProvider),
     findSelectorLevelButtons: function (level: number): WrapperArray<Vue> {
       return findTestDataById(wrapper, 'variants-list')
         .at(level)
@@ -92,12 +92,12 @@ describe('results with variants', () => {
       <span data-test="result-name">{{newResult.name}}</span>
     `;
     const result = createResultStub('tshirt');
-    const { wrapper } = renderVariantsResultProvider({ result, template });
+    const { wrapper } = renderResultVariantsProvider({ result, template });
     expect(wrapper.find(getDataTestSelector('result-name')).text()).toBe('tshirt');
   });
 
   it('merges the selected and parent variants data with the result', async () => {
-    const { wrapper, findSelectorLevelButtons } = renderVariantsResultProvider({
+    const { wrapper, findSelectorLevelButtons } = renderResultVariantsProvider({
       template: `
         <div>
           <ResultVariantSelector #variant-content="{variant}">
@@ -143,7 +143,7 @@ describe('results with variants', () => {
   });
 
   it('keeps the original result unmodified', async () => {
-    const { wrapper } = renderVariantsResultProvider({
+    const { wrapper } = renderResultVariantsProvider({
       template: `
         <div>
           <ResultVariantSelector/>
@@ -161,7 +161,7 @@ describe('results with variants', () => {
   });
 
   it('emits UserSelectedAResultVariant event when a variant is selected', async () => {
-    const { wrapper } = renderVariantsResultProvider({ result });
+    const { wrapper } = renderResultVariantsProvider({ result });
     const eventsSpy = jest.spyOn(XPlugin.bus, 'emit');
 
     const button = wrapper.find(getDataTestSelector('variant-button'));
@@ -182,14 +182,14 @@ describe('results with variants', () => {
 
   describe('result variant selector', () => {
     it('renders the whole variant by default', () => {
-      const { wrapper } = renderVariantsResultProvider({ result });
+      const { wrapper } = renderResultVariantsProvider({ result });
       const button = wrapper.find(getDataTestSelector('variant-button'));
       expect(JSON.parse(button.text())).toEqual(variants[0]);
     });
 
     it('add selected class when a variant is selected', async () => {
       const className = 'x-result-variant-selector__item--is-selected';
-      const { wrapper } = renderVariantsResultProvider({ result });
+      const { wrapper } = renderResultVariantsProvider({ result });
 
       const firstVariantButton = wrapper.find(getDataTestSelector('variant-button'));
       const variantWrappers = wrapper.findAll(getDataTestSelector('variant-item'));
@@ -219,7 +219,7 @@ describe('results with variants', () => {
         </div>
       `;
 
-      const { findSelectorLevelButtons } = renderVariantsResultProvider({
+      const { findSelectorLevelButtons } = renderResultVariantsProvider({
         template,
         result
       });
@@ -249,7 +249,7 @@ describe('results with variants', () => {
     });
 
     it('wont render if no result is injected', () => {
-      const { wrapper } = renderVariantsResultProvider({
+      const { wrapper } = renderResultVariantsProvider({
         result: null
       });
 
@@ -257,7 +257,7 @@ describe('results with variants', () => {
     });
 
     it('wont render if the provided result does not have variants', () => {
-      const { wrapper } = renderVariantsResultProvider({
+      const { wrapper } = renderResultVariantsProvider({
         result: createResultStub('jacket')
       });
 
@@ -265,7 +265,7 @@ describe('results with variants', () => {
     });
 
     it('exposes variants, selectedVariant and selectVariant in the default slot', async () => {
-      const { wrapper } = renderVariantsResultProvider({
+      const { wrapper } = renderResultVariantsProvider({
         template: `
           <ResultVariantSelector #default="{variants, selectedVariant, selectVariant}" >
             <div>
@@ -299,7 +299,7 @@ describe('results with variants', () => {
     });
 
     it('exposes variant, isSelected and selectVariant in the variant slot', async () => {
-      const { wrapper } = renderVariantsResultProvider({
+      const { wrapper } = renderResultVariantsProvider({
         template: `
           <ResultVariantSelector #variant="{variant, selectVariant, isSelected}">
             <button
@@ -325,7 +325,7 @@ describe('results with variants', () => {
     });
 
     it('exposes variant and isSelected in the variant-content slot', async () => {
-      const { wrapper } = renderVariantsResultProvider({
+      const { wrapper } = renderResultVariantsProvider({
         template: `
           <ResultVariantSelector #variant-content="{variant, isSelected}">
             {{variant.name}}<span v-if="isSelected"> SELECTED!</span>
@@ -347,9 +347,9 @@ describe('results with variants', () => {
 });
 
 /**
- * The options for the `renderVariantsResultProvider` function.
+ * The options for the `renderResultVariantsProvider` function.
  */
-interface VariantsResultProviderOptions {
+interface ResultVariantsProviderOptions {
   /** The result containing the variants. */
   result: Result | null;
   /** The template to render inside the provider's default slot. */
@@ -357,10 +357,10 @@ interface VariantsResultProviderOptions {
 }
 
 /**
- * Test API for the {@link VariantsResultProvider} component.
+ * Test API for the {@link ResultVariantsProvider} component.
  */
-interface VariantsResultProviderApi {
-  /** The wrapper for {@link VariantsResultProvider} component. */
+interface ResultVariantsProviderApi {
+  /** The wrapper for {@link ResultVariantsProvider} component. */
   wrapper: Wrapper<Vue>;
   /**
    * Util function to find the variant buttons of a level.
